Session avec Smaine résolution de divers petit bug, tour d'horizon du site et des dernieres modifications a éffectuer, mise en place d'une propriété css  word-wrap afin que le texte force le retour a la ligne pour le contenu des chapitres, mise en place d'un niveau d'administration pour les personnes inscrit ou non

+++ html/view/frontend/lastThreePostsView.php
@@ -27,7 +27,9 @@
 
     -
 
-    <input class="btn btn-danger buttonCenter" type="button" value="Supprimer" onclick="javascript:location.href='index.php?action=deletePost&id=<?= $post->getId(); ?>'">
+    <?php if($_SESSION['admin']): ?>
+        <input class="btn btn-danger buttonCenter" type="button" value="Supprimer" onclick="javascript:location.href='index.php?action=deletePost&id=<?= $post->getId(); ?>'">
+    <?php endif; ?>
 
 <hr>
   <?php endforeach; ?>
